Remove "awk" and "paste" and "tail" command, and use "sed" command, to avoid unintended buffering
On some environment, such as debian, the output of this script seems to stop for several seconds.
It might be caused by buffering feature of "awk" or "paste" or "tail" command.
So replace them to "sed" command. This version may generate the output in every second, even in such environment.